requirements.txt
langchain==0.3.27
langchain_community==0.3.29
langchain_openai==0.3.33
langchain_pinecone==0.2.12
# langchain-pinecone==0.1.0
pinecone==7.3.0
python-dotenv==1.1.1
streamlit==1.47.1
pymupdf
